+++ design-toolkit/components/src/components/combobox/index.tsx
@@ -15,18 +15,26 @@ import {
   Button as AriaButton,
   ComboBox as AriaComboBox,
   type ComboBoxProps as AriaComboBoxProps,
+  Header as AriaHeader,
   Input as AriaInput,
   type InputProps as AriaInputProps,
   ListBox as AriaListBox,
+  Collection as AriaListBoxCollection,
   ListBoxItem as AriaListBoxItem,
   type ListBoxItemProps as AriaListBoxItemProps,
+  ListBoxSection as AriaListBoxSection,
+  type ListBoxSectionProps as AriaListBoxSectionProps,
+  ListLayout as AriaListLayout,
+  type ListLayoutOptions as AriaListLayoutOptions,
   Popover as AriaPopover,
   Text as AriaText,
+  Virtualizer as AriaVirtualizer,
+  type VirtualizerProps as AriaVirtualizerProps,
 } from 'react-aria-components';
 
 import { cn } from '@/lib/utils';
 import type { MenuItem } from '@/types/types';
-import { ChevronDown } from '@accelint/icons';
+import ChevronDown from '@accelint/icons/chevron-down';
 import { type VariantProps, cva } from 'cva';
 import type { ReactNode } from 'react';
 import { Icon } from '../icon';
@@ -100,7 +108,8 @@ export interface ComboBoxProps<T extends MenuItem>
       VariantProps<typeof textFieldStyles>,
       'isDisabled' | 'isInvalid' | 'isReadOnly'
     >,
-    Omit<AriaComboBoxProps<T>, 'children'> {
+    Omit<AriaComboBoxProps<T>, 'children'>,
+    Pick<AriaVirtualizerProps<AriaListLayoutOptions>, 'layoutOptions'> {
   className?: string;
   children: ReactNode | ((item: T) => ReactNode);
   description?: string;
@@ -119,6 +128,7 @@ export function ComboBox<T extends MenuItem>({
   isReadOnly,
   label,
   placeholder,
+  layoutOptions,
   size = 'medium',
   ...props
 }: ComboBoxProps<T>) {
@@ -160,7 +170,7 @@ export function ComboBox<T extends MenuItem>({
                 isReadOnly && 'hidden',
               ])}
             >
-              <Icon>
+              <Icon className='block'>
                 <ChevronDown />
               </Icon>
             </AriaButton>
@@ -185,10 +195,15 @@ export function ComboBox<T extends MenuItem>({
             </AriaText>
           )}
           <AriaPopover className='w-(--trigger-width)'>
-            <AriaListBox className='grid max-h-[200px] grid-cols-[auto_1fr] overflow-y-auto overflow-x-clip rounded-medium bg-surface-overlay shadow-elevation-overlay outline outline-static-light'>
-              {/* @ts-expect-error package version mismatch TODO */}
-              {children}
-            </AriaListBox>
+            <AriaVirtualizer
+              layout={AriaListLayout}
+              layoutOptions={layoutOptions}
+            >
+              <AriaListBox className='grid max-h-[200px] grid-cols-[auto_1fr] overflow-y-auto overflow-x-clip rounded-medium bg-surface-overlay shadow-elevation-overlay outline outline-static-light'>
+                {/* @ts-expect-error package version mismatch TODO */}
+                {children}
+              </AriaListBox>
+            </AriaVirtualizer>
           </AriaPopover>
         </>
       )}
@@ -226,14 +241,18 @@ function ComboBoxItem<T extends MenuItem>({
       {...props}
       className={comboBoxItemStyles({ className })}
     >
-      {(props) => {
-        if (children) {
-          return typeof children === 'function' ? children(props) : children;
+      {(renderProps) => {
+        if (typeof children === 'function') {
+          return children(renderProps);
         }
 
         return (
           <>
-            {icon && <span className='mr-xs'>{icon}</span>}
+            {icon && (
+              <span className='mr-xs'>
+                <Icon>{icon}</Icon>
+              </span>
+            )}
             <div className='col-start-2 flex flex-col gap-xxs'>
               <AriaText slot='label'>{name}</AriaText>
               {description && (
@@ -251,3 +270,28 @@ function ComboBoxItem<T extends MenuItem>({
 ComboBoxItem.displayName = 'ComboBox.Item';
 
 ComboBox.Item = ComboBoxItem;
+
+interface ComboBoxSectionProps<T extends MenuItem>
+  extends AriaListBoxSectionProps<T> {
+  header?: string;
+}
+
+export function ComboBoxSection<T extends MenuItem>({
+  children,
+  header,
+  items,
+}: ComboBoxSectionProps<T>) {
+  return (
+    <AriaListBoxSection
+      id={header}
+      className='col-span-2 mt-s grid grid-cols-[auto_1fr] border-default-dark border-t first:border-none'
+    >
+      <AriaHeader className='col-span-2 m-xs my-s text-default-dark text-header-xs'>
+        {header}
+      </AriaHeader>
+      <AriaListBoxCollection items={items}>{children}</AriaListBoxCollection>
+    </AriaListBoxSection>
+  );
+}
+ComboBox.displayName = 'ComboBox.Section';
+ComboBox.Section = ComboBoxSection;